Only filter/weigh hosts once if scheduling a single instance
If we're only scheduling for a single instance, we don't need
filter and weigh hosts another time before picking alternates
because the single selected host for the single instance can't
be in the list of filtered/weighed alternates anyway.

This helps avoid an unnecessary performance penalty for finding
alternates when we just have a single instance - which is likely
the majority of cases.

Change-Id: Ie065c4e169db4e02fbfa8ccadb0af23d2784cab8