Fix AttributeError in BlockDeviceMapping.obj_load_attr
The BDM has no uuid attribute so the debug message in here
would result in an AttributeError. This has been around since
the creation of this object, and the debug log message was
probably copied from the Instance object.

This was only exposed in Pike when this code started
lazy-loading the instance field:

  I1dc54a38f02bb48921bcbc4c2fdcc2c946e783c1

So this change fixes that bug and adds tests for obj_load_attr.

Change-Id: I8b55227b1530a76c2f396c035384abd89237d936
Closes-Bug: #1726871
(cherry picked from commit 1ca191fcc4d809c991c23dedc951bbe7206edf1d)