Accept all standard resource classes in flavor extra specs
When transforming flavor resources into resource classes, we look at any
'resources' extraspec key to know which resource classes to ask.
That said, we were previously only accepting custom resource classes or
existing standard resource classes if and only if they were in the flavor
(and then overriding them).
Given operators would want to ask for specific resource classes (like VGPU)
that are not flavor fields, we need to modify the utils method to make sure
we accept them.

Partially Implements: blueprint add-support-for-vgpu
Co-Authored-By: Huan Xie <huan.xie@citrix.com>

Change-Id: I84ea7779b196ec535f4109c5343c8de9056a2b2d