Add VIFMigrateData object for live migration
- This change adds a new VIFMigrateData object.
- This change adds a new vifs field to the
  base LiveMigrateData object.

Later patches will populate this information during
live migration where nova uses the neutron port
binding extended API to bind the port on the destination
host, and store the destination host binding details
in this object to be leveraged on the source host for
things like defining vif details in the guest before
starting the live migration in the hypervisor.

Note that vif_details and profile must be stored as
serialized dicts similar to
LibvirtLiveMigrateBDMInfo.connection_info_json.

Co-Authored-By: Matt Riedemann <mriedem.os@gmail.com>

Part of blueprint neutron-new-port-binding-api

Change-Id: I174ae21f143016bb23b8043fa22465dac6c71314