Remove incorrect comment about instance.locked
The comment about instance.locked is deprecated
seems incorrect and missleading. As instance.locked
is still used in check:
https://github.com/openstack/nova/blob/16.0.0.0rc2/nova/compute/api.py#L195

This is because that drop the field needs major verison
bump on instance obj and dropping the column from DB with
offline data migration, and it is probably what we never
do because instances table is a very huge table, it could
be too expensive to do this.

Change-Id: Iefd0ae1d66178899c7cd44b5336b3dc6a18496a8