Use the RequestSpec when getting scheduler_hints in compute
For the first chosen host during an instance build, conductor
will pass both filter_properties and request_spec and they
both have scheduler_hints.

We want to eventually get rid of the legacy filter_properties,
so this change uses the request_spec, if available, to get the
scheduler_hints in the compute during the build. These are only
used when validating server group (anti-)affinity policies.

If the request_spec isn't available, because we rescheduled to
another host for example, then we fallback to the filter_properties
to get the hints.

Part of blueprint request-spec-use-by-compute

Change-Id: I49ffebcd129990f1835f404d98b51732a32171eb