Pass RequestSpec to ConductorTaskAPI.build_instances
There are two things that call the conductor task API's
build_instances method:

1. nova-api when using cellsv1
2. nova-compute when rescheduling during a build failure

Since we want to eventually replace the filter_properties
usage in the compute service with the request_spec, this
allows passing the request_spec from the compute service
during a reschedule.

For now we have to continue munging that RequestSpec a
bit using the filter_properties for retries and the
next chosen host from the scheduler, but that can eventually
go away. With this, the computes can start to at least
pull scheduler_hints out of the request_spec rather than
the filter_properties, which is one more step in removing
the compute's dependency on filter_properties.

Note that we don't pass the request_spec from nova-api
for cells v1, simply because cells v1 is deprecated.

Part of blueprint request-spec-use-by-compute

Change-Id: Ie5233bd481013413f12e55201588d37a9688ae78