begin refactor AllocCandidates._get_by_filters()
The AllocationCandidates._get_by_filters() method is a long, complex
mega-method. We want to detangle it and make the code in it
understandable and readable. The first step in this refactoring involves
renaming a few poorly-named variables, adding more code comments, and
breaking out the SQL for the most common code path (where no providers
share resources with other providers) into a separate module-level
function.

Change-Id: I716b279f719eb1fe26401a776344b976eca6587a