placement: Contributor doc microversion checklist
I put out a new placement microversion [1] and initially missed updating
the API reference documentation [2].  I'd like to say it would have
helped if the checklist in the Microversion section of the contributor
doc had been up to date, except I didn't even know about that document
at the time.  Anyway, this change set adds that bullet - and I'll know
to refer to it next time.

[1] https://review.openstack.org/499826
[2] https://review.openstack.org/515748

Change-Id: Id0329d9824eeb1d210defe0567286372b72cb1cf