Raise specific exception when swapping migration allocations fails
If we fail to find resource allocations for an instance on the
source compute node when swapping allocations for a cold migrate /
resize operation, we need to raise a specific exception which will
eventually be a 500 internal error from the REST API. Using
InstanceUnacceptable is incorrect since it extends Invalid which is
handled in the resize REST API controller code and returned as a 400
bad request with the message "Instance image invalid" which is not
at all the actual failure.

As for the actual reason why the instance allocation on the source
node is not found, that is a different bug.

Change-Id: Id8e2dcf21776f8237a8b63acb23787bb42c9bd13
Closes-Bug: #1729356