handle traits with sharing providers
Filtering allocation candidates on a set of required traits when there
are sharing providers in the deployment adds a lot of complexity. This
is just a simple fact of life, unfortunately. We have to play a sort of
game of scheduling Tetris whereby after we have grabbed the list of
resource providers that have access to local or shared inventory for
requested resources, we then need to ensure that the group of resource
providers that comprise the allocation request have ALL of the required
traits collectively between them.

This patch adds support for filtering allocation candidates by a set of
required traits when there are sharing providers in the deployment. The
code could still use additional cleanup, in particular there are areas
in the (now even larger) _alloc_candidates_with_shared() function that
do a lookup of a resource provider internal ID by traversing the list of
resource provider summaries looking for a resource provider by UUID.
These areas of the function should be refactored and a UUID -> ID lookup
hash created to remove this duplicate code.

Change-Id: Ia2295414654abf0e282926be6cd068189cd75352