Fix return type in FilterScheduler._legacy_find_hosts
The FilterScheduler._schedule method should be returning
a list of list of selected hosts. When include_alternatives
is False in _legacy_find_hosts, it was only returning back
a list of hosts, which would result in an IndexError when
select_destinations() tries to take the first entry from each
item in the list.

Change-Id: Ia6c87900605d3604beb74b942b0e30575b814112
Closes-Bug: #1729445