Remove instance.keypairs migration code
This was all added in Newton to prepare for moving
keypairs to the API database. The 345 nova DB migration
in Ocata forced people to run the online data migrations
to move keypairs out of the nova (cell) database to the
nova_api database. When they ran those data migrations,
they'd also be migrating instance.keypairs into the
instance_extra table.

At this point we can remove all of this migration code.

Change-Id: Ie83e7bd807c2c79e5cbe1337292c2d1989d4ac03