Avoid redundant security group queries in GET /servers/{id}/os-security-groups
The GET /servers/{server_id}/os-security-groups API code can
perform poorly if the instance has several security groups and
each security group has several rules. This is because when processing
the output, we loop over the groups, and loop over the rules per group,
and then for each rule, if it has a group_id specified, we query
the security group details (from Neutron in most cases).

If more than one rule points at the same group_id, we're doing a redundant
group lookup and sending more traffic to the security group API (aka Neutron)
than needed.

This change optimizes that single API to load the rule group details
up front so that we only do at most one lookup per group_id.

This could be extended to GET /os-security-groups but that API is
deprecated so any optimization there is lower priority.

Change-Id: Ia451429f61b15526fade6838386e562c17591d36
Closes-Bug: #1729741