ironic: Report resources as reserved when needed
The main use case for this is during ironic node cleaning, when,
instead of deleting the resource provider from placement we can
just report the resources as temporarily reserved.

This change also bumps placement API microversion used to update
inventory to '1.26' which allows to have inventories with
reserved value equal to total.

Closes-Bug: 1771577
blueprint: allow-reserved-equal-total-inventory
Change-Id: I1fd85860c96e8690fbcf93c8a2f02178168bfd5a