Refactor test_allocation_candidates
For the sake of code deduplication, LOC reduction, and readability
(improved signal:noise ratio), this change set implements some
convenience methods in AllocationCandidatesTestCase for creation of
placement artifacts such as resource providers, inventory, and
allocations, as well as for extracting and comparing results of
get_allocation_candidates.

This only changes test code; and is just a refactor - no tests are
added, removed, or substantively changed.

Change-Id: I34527a3b2ac33ec3b1c6cce2c4a2163eea1ada4d