rt: Make resource tracker always invoking get_inventory()
This is a stable/pike only backport from queens. This change originally
was made of part of the feature bp add-support-for-vgpu. Bit it is
actually fixing a bug in the resource tracker and not adding any VGPU
specific code. I amended the commit message and the code comments to
adapt it to pike where  VGPU is not a thing.

The ironic custom resource stat is not saved in the compute_nodes.
Instead it's reported by drivers' get_inventory() interface and the
inventory data is saved via placement. But the resource tracker will
skip invoking get_inventory() if the resource in compute_node is not
changed. It will cause the ironic customer resource is not added to
placement if other compute node resources are not changed.

This commit is to change resource tracker to always proceed to invoke
the function of get_inventory and use scheduler_client interfaces to
update inventory to placement. The scheduler_client will ensure the
update request to placement only happen when inventory is changed
via comparing with the local cached inventory.

Change-Id: I6e204fe8e7c003246c9d8bebf484323700737093
(cherry picked from commit e2a18a37190e4c7b7697a8811553d331e208182c)