rt: Make resource tracker always invoking get_inventory()
The VGPU resource stat is not saved in the compute_nodes. Instead it's
reported by drivers' get_inventory() interface and the inventory data
is saved via placement. But the resource tracker will skip invoking
get_inventory() if the resource in compute_node is not changed. It
will cause the VGPU resource cannot be updated if other compute node
resources are not changed.

This commit is to change resource tracker to always proceed to invoke
the function of get_inventory and use scheduler_client interfaces to
update inventory to placement. The scheduler_client will ensure the
update request to placement only happen when inventory is changed
via comparing with the local cached inventory.

Change-Id: I6e204fe8e7c003246c9d8bebf484323700737093
blueprint: add-support-for-vgpu