Start moving users of parted to privsep.
This isn't all calls, but a step along the way. Note that I moved
a xenapi unit test to being the first unit test for privsep directly.
I'm not sure if I love this way of testing (having the "meat" of
the method undecorated and having a wrapper), but I've been unable
to come up with a way to mock away the entrypoint wrapper.

Change-Id: I35d4e7c6c43cbcbc1549aa28796ff97f447a4b03
blueprint: hurrah-for-privsep