unquiesce instance on volume snapshot failure
This patch adds an exception catch to "snapshot_volume_backed()" of
compute/api.py that catches (at the moment) _all_ exceptions from the
underlying cinderclient. Previously, if the instance is quiesced ( frozen
filesystem ) then the exception will break execution of the function,
skipping the needed unquiesce, and leave the instance in a frozen state.

Now, the exception catch will unquiesce the instance if it was prior to
the failure.

Got a unit test in place with the help of Matt Riedemann.
    test_snapshot_volume_backed_with_quiesce_create_snap_fails

Change-Id: I60de179c72eede6746696f29462ee9d805dace47
Closes-bug: #1731986