Test alloc_cands with one RP shared between two RPs
This patch proposes two new test cases that tries to cover the cases
when a sharing RP is connected to two non sharing RPs.

Then the first case only differs from the previously covered scenario
that is connects the two CNs to the sharing RP via two different
aggregates:

    CN1 (VCPU)        CN2 (VCPU)
        \ agg1        / agg2
         SS1 (DISK_GB)

This works as expected.

Then in the third case both CNs will be connected to an additional,
separate, sharing RP in the following way:

       CN1 (VCPU)          CN2 (VCPU)
      / agg3   \ agg1     / agg1   \ agg2
  SS3 (IPV4)   SS1 (DISK_GB)      SS2 (IPV4)

Then VCPU, DISK_GB and IPV4 resources are requested. However in this
case besides the expected candidates we get invalid candidates mixing
CN1 with SS2 and CN2 with SS3.

Related-Bug: #1732707
Change-Id: I2c815ef0f10120735051429f5c290a49d8d1b11e