Refactor the code to check for sufficient hosts
There are multiple places in the filter_scheduler code where we need to
check that we found enough hosts to match the request. This patch
refactors the code to create a single method that checks for that, and
raises a NoValidHost exception if there are not enough hosts.

Blueprint: return-alternate-hosts

Change-Id: I858e7f19950c12421331cc2fbb033cd83fb0e260