Add regression test for rebuild with new image doubling allocations
Commit 984dd8ad6add4523d93c7ce5a666a32233e02e34 makes a rebuild
with a new image go through the scheduler again to validate the
image against the instance.host (we rebuild to the same host that
the instance already lives on).

The problem is that change introduced a regression where the
FilterScheduler is going to think it's doing a resize to the same
host and double the allocations for the instance (and usage for the
compute node provider) in Placement, which is wrong since the
flavor is the same.

This adds a regression test to show the bug.

NOTE(mriedem): Due to cc833359870d3962326c35094adea2f525ec8141 not
being in Pike, we have to use the MediumFakeDriver until the bug
is fixed. Also, the ComputeManager._get_nodename method doesn't exist
in Pike so we have to just get the hypervisor uuid from the os-hypervisors
API.

Change-Id: Ie0949b4e6101f0b29ec4542146d523a07a683991
Related-Bug: #1664931
(cherry picked from commit cacfd372acb4eb056c4391db3c988bfe91c957df)